test: shared conftest+markers, dedupe fixtures, drop no-op/dup tests
- add tests/conftest.py: datamaxi/telegram/naver fixtures, API_KEY/BASE_URL,
  _FLAKY_PROD_DATA_XFAIL; remove copy-paste from test_call/test_integration
- register integration/smoke markers; --strict-markers; drop unused errors
- mark live lanes (integration/smoke) so keyless CI deselects via -m
- drop no-op TestErrorHandling + duplicate test_candle_with_from_unix
