+++ setup.cfg
@@ -2,6 +2,7 @@
 description_file = README.md
 
 [tool:pytest]
+addopts = --strict-markers
 markers =
     cex: tests for CEX endpoints
     funding: tests for funding rate endpoints
@@ -9,8 +10,9 @@ markers =
     forex: tests for forex endpoints
     telegram: tests for telegram endpoints
     naver: tests for naver endpoints
-    errors: tests for error handling
     types: tests for response types
+    integration: live tests that hit prod endpoints (need DATAMAXI_API_KEY)
+    smoke: live alive-check lane pinging each endpoint once (need DATAMAXI_API_KEY)
 
 [flake8]
 exclude =
